Create README.md
Q:uiz is a Web-Application, in Italian, created to test yourself with a multiple-choice quiz on general culture. For each correct answer the user will be assigned a point. At the end of the game the player will be shown his answers, the correct answer and the score.
For the creation of the site I used Vite with React and Typescript.
For styling I used Tailwind and Css3.
For deploying the site I utilized Netlify.